Add files via upload
This is used to connect the users to the MySQL database. Make sure this folder goes into htdocs as it's the only way this works. 

Open XAMPP and start MySQL

click on admin, and it opens a page in your browser 

In phpMyAdmin (left sidebar), click New

Under “Create database”, enter:

myapp


Choose utf8_general_ci for collation

Click Create